requirements.txt
sentence-transformers
numpy<2.0
scipy
pandas
matplotlib
seaborn
plotly
dash
dash-bootstrap-components
scikit-learn
umap-learn
pyyaml
pytest
pytest-cov
pytest-mock
jupyterlab
ipywidgets

